pyproject.toml
[project]
name = "zorvyn-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]
